Fix for active true mutations not being saved
When the player gets permanently TFed he or she gets some true mutations, but the setting wasn't stored for the player and therefore not stored in the save game.
Instead the property was set to true in the IMutation-class itself. This led to two issues:
First off, when you get a true mutation, like for example Twin Heart from the sand worm permTF and then load a save where you are not permanently TFed, you'll still have the true mutation active, so I still got that mutation without Evangeline when switching back to my dragon-taur.
Then, when you close the game client and load it again later, the setting is reverted to false.