Reintroduced changeLimit-managment to demon TFs
In the previous PR I've cleaned up demon changes because the changeLimit management wasn't implemented correctly there.

This time I've reintroduced it and made it consistent now and it now adds to the TIMES_TRANSFORMED-flag as intended.

Side note: The change to mouse cocoa were for my first tests, if using the global instead of the local vars works as intended and it did. I've left that change intact because it doesn't affect gameplay in any way.